backend/services/video_analysis.py
"""
ビデオ分析サービス (表情・視線・姿勢)
Phase 4 で実装予定
"""
import cv2
import numpy as np
from typing import Dict, Any, List
import logging

logger = logging.getLogger(__name__)

# Phase 4 で実装
# import mediapipe as mp


def analyze_video(video_path: str) -> Dict[str, Any]:
    """
    ビデオから表情・視線・姿勢を分析 (Phase 4)
    
    Args:
        video_path: ビデオファイルパス
    
    Returns:
        {
            "facial_expression": {...},  # 表情分析
            "eye_contact": {...},         # 視線追跡
            "posture": {...},             # 姿勢分析
            "gestures": {...}             # ジェスチャー検出
        }
    """
    logger.warning("ビデオ分析は Phase 4 で実装予定です")
    
    return {
        "facial_expression": _analyze_facial_expression_stub(),
        "eye_contact": _analyze_eye_contact_stub(),
        "posture": _analyze_posture_stub(),
        "gestures": _analyze_gestures_stub(),
        "status": "Phase 4 で実装予定"
    }


def _analyze_facial_expression_stub() -> Dict[str, Any]:
    """
    表情分析 (スタブ)
    
    実装予定:
    - MediaPipe Face Mesh で顔のランドマーク検出
    - 笑顔・緊張・困惑などの感情分類
    - 表情の変化を時系列で記録
    """
    return {
        "smile_frequency": 0,  # 笑顔の頻度 (回/分)
        "smile_duration": 0,   # 笑顔の持続時間 (秒)
        "tension_level": 0,    # 緊張レベル (0-100)
        "dominant_emotion": "neutral",  # 主な表情
        "emotion_timeline": []  # タイムスタンプ付き表情変化
    }


def _analyze_eye_contact_stub() -> Dict[str, Any]:
    """
    視線追跡 (スタブ)
    
    実装予定:
    - MediaPipe Face Detection で視線方向を推定
    - カメラ (面接官) を見ている時間を計測
    - 視線の逸らし方のパターン分析
    """
    return {
        "eye_contact_ratio": 0,      # アイコンタクト比率 (%)
        "eye_contact_frequency": 0,   # 視線を合わせる頻度 (回/分)
        "average_duration": 0,        # 平均持続時間 (秒)
        "looking_away_patterns": [],  # 視線を逸らすパターン
        "feedback": "カメラを見る時間を増やしましょう"
    }


def _analyze_posture_stub() -> Dict[str, Any]:
    """
    姿勢分析 (スタブ)
    
    実装予定:
    - MediaPipe Pose で身体のランドマーク検出
    - 猫背・前傾姿勢を検出
    - 姿勢の変化を時系列で記録
    """
    return {
        "posture_score": 0,          # 姿勢スコア (0-100)
        "slouching_detected": False,  # 猫背検出
        "leaning_forward": False,     # 前傾姿勢
        "body_stability": 0,          # 身体の安定性
        "posture_changes": [],        # 姿勢変化の記録
        "feedback": "背筋を伸ばして座りましょう"
    }


def _analyze_gestures_stub() -> Dict[str, Any]:
    """
    ジェスチャー検出 (スタブ)
    
    実装予定:
    - MediaPipe Hands で手の動きを追跡
    - 手を組む・髪を触るなどの癖を検出
    - ジェスチャーの種類と頻度を記録
    """
    return {
        "hand_gestures": [],           # 検出されたジェスチャー
        "fidgeting_count": 0,          # そわそわした動き (回)
        "hair_touching": 0,            # 髪を触る回数
        "hand_clasping": 0,            # 手を組む回数
        "inappropriate_gestures": [],  # 不適切なジェスチャー
        "feedback": "手の動きは自然です"
    }


# === Phase 4 で実装する関数 (TODO) ===

def analyze_facial_expression_mediapipe(video_path: str) -> Dict[str, Any]:
    """
    MediaPipe Face Mesh で表情分析
    
    実装手順:
    1. pip install mediapipe opencv-python
    2. mp_face_mesh = mp.solutions.face_mesh.FaceMesh()
    3. ビデオフレームごとにランドマーク検出
    4. 口角・眉の位置から笑顔・緊張を判定
    """
    pass


def analyze_eye_contact_mediapipe(video_path: str) -> Dict[str, Any]:
    """
    MediaPipe Face Detection で視線追跡
    
    実装手順:
    1. mp_face_detection = mp.solutions.face_detection.FaceDetection()
    2. 顔の向きと目の位置を検出
    3. カメラ方向を向いている時間を計測
    """
    pass


def analyze_posture_mediapipe(video_path: str) -> Dict[str, Any]:
    """
    MediaPipe Pose で姿勢分析
    
    実装手順:
    1. mp_pose = mp.solutions.pose.Pose()
    2. 肩・背中のランドマークを検出
    3. 角度を計算して猫背・前傾を判定
    """
    pass


def analyze_gestures_mediapipe(video_path: str) -> Dict[str, Any]:
    """
    MediaPipe Hands でジェスチャー検出
    
    実装手順:
    1. mp_hands = mp.solutions.hands.Hands()
    2. 手のランドマークを追跡
    3. 動きのパターンから癖を検出
    """
    pass
