requirements.txt
# === Core Framework ===
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0
websockets>=12.0  # リアルタイム進捗通知

# === Template & Report ===
jinja2>=3.1.2
markdown>=3.5.0  # Markdown プレビュー用

# === Audio Processing ===
faster-whisper>=0.9.0
librosa>=0.10.1
soundfile>=0.12.1
numpy>=1.24.0
scipy>=1.11.0

# === AI Analysis ===
google-generativeai>=0.3.0

# === Utilities ===
python-dateutil>=2.8.0
requests>=2.31.0

# === Development Tools ===
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # FastAPI TestClient 用

# === Video Analysis (Phase 4) ===
opencv-python>=4.8.0
mediapipe>=0.10.0

# === 将来的な拡張 (Phase 2+) ===
# weasyprint>=60.0  # PDF 生成
# SQLAlchemy>=2.0.0  # DB 管理
# alembic>=1.12.0  # DB マイグレーション
# pandas>=2.1.0  # データ分析
